requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
requests>=2.25.0
